requirements.txt
xgboost
kafka-python
psycopg2-binary
redis
pandas
numpy
scikit-learn
joblib
prometheus-client
python-snappy